+++ include/mettle/test_uid.hpp
@@ -9,7 +9,7 @@ namespace mettle {
   using test_uid = std::uint64_t;
 
   namespace detail {
-    inline std::atomic<test_uid> next_test_uid(0);
+    inline std::atomic<test_uid> next_test_uid(1);
     inline test_uid make_test_uid() { return next_test_uid++; }
   }